pyproject.toml
[project]
name = "restaurant-recommendation-m1"
version = "0.1.0"
description = "Phase 1: Zomato HF dataset → canonical catalog (Parquet)"
requires-python = ">=3.11"
dependencies = [
  "datasets>=2.16.0",
  "fastapi>=0.109.0",
  "groq>=0.4.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.0",
  "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = ["httpx>=0.26.0", "pytest>=7.0.0"]
streamlit = ["streamlit>=1.28.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
